requirements.txt
pycryptodome==3.16.0
pyperclip==1.8.2
mysql-connector-pyton==8.0.13
pyotp==2.8.0